requirements.txt
numpy
pyvista
pillow